修正: 時計表示の幅計算とコロン描画の位置を調整
- 時計表示の全体幅計算において、コロンの幅を修正しました。
- コロン描画関数内でのピクセル位置計算を簡素化しました。

+++ src/font.rs
@@ -244,7 +244,7 @@ pub fn draw_colon(
             if (bits >> (1 - col)) & 1 == 1 {
                 draw_pixel_block(
                     fb,
-                    x + (col as i32) * pixel_size,
+                    x + col * pixel_size,
                     y + (row as i32) * pixel_size,
                     pixel_size,
                     color,
